refactor: RootStackParamList의 패키지 이동
