+++ fourtoon-cookie/src/pages/DiaryTimelinePage/DiaryComponent/Header/Header.tsx
@@ -5,7 +5,7 @@ import DOTS_ICON from "../../../../../assets/icon/dots.png";
 import { useActionSheet } from "@expo/react-native-action-sheet";
 import { useSelectedCharacterStore } from "../../../../hooks/store/selectedCharacter";
 import { NavigationProp, useNavigation } from "@react-navigation/native";
-import { RootStackParamList } from "../../../../constants/routing";
+import { RootStackParamList } from "../../../../types/routing";
 import handleError from "../../../../error/errorhandler";
 import { GlobalErrorInfoType } from "../../../../types/error";
 import { useDeleteDiary } from "../../../../hooks/server/diary";
