+++ fourtoon-cookie/src/pages/DiaryWritePage/WriteDoneButtonLayout/WriteDoneButtonLayout.tsx
@@ -5,7 +5,7 @@ import { NavigationProp, useNavigation } from "@react-navigation/native";
 import { OS } from "../../../types/os"
 import { Diary, DiaryStatus } from "../../../types/diary";
 import { useSelectedCharacterStore } from "../../../hooks/store/selectedCharacter";
-import { RootStackParamList } from "../../../constants/routing";
+import { RootStackParamList } from "../../../types/routing";
 import { API_STATUS } from "../../../types/api";
 import { ApiError } from "../../../types/error/ApiError";
 import handleError from "../../../error/errorhandler";
